+++ docs/Changelog.rst
@@ -3,6 +3,19 @@
 Changelog
 =========
 
+8.6.2.post1 (23 March 2023)
+---------------------------
+
+**Announcement**
+
+  Late 8.6.2 post-release.
+
+**Changes**
+
+  * Fix certificate detection. See #33.
+  * Add one-line API Key support (the Base64 encoded one).
+  * Update docs to reflect base64 token API Key functionality.
+
 8.6.2 (19 February 2023)
 ------------------------